DTSBPS-611: code optimisation
